requirements.txt
absl-py==1.2.0
asttokens==2.0.5
astunparse==1.6.3
autopep8==1.6.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
flatbuffers==1.12
fonttools==4.33.2
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.21.1
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
stack-data==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tk==0.1.0
toml==0.10.2
tornado==6.1
traitlets==5.2.1.post0
typing_extensions==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
